Relocate Stickers component
