+++ src/components/Header/Header.astro
@@ -99,6 +99,7 @@ const topLevelPath = currentPath.split("/")[1];
     display: flex;
     align-items: stretch;
     gap: var(--space-2xs);
+    margin-inline-end: calc(-1 * var(--space-s));
 
     a,
     button {
