--- src/components/Toys/Toys.astro
@@ -1,17 +1,0 @@
----
-import { Stickers } from "./Stickers/Stickers";
----
-
-<div class="toys" transition:persist>
-  <Stickers client:idle />
-</div>
-
-<style>
-  .toys {
-    position: fixed;
-    inset: 0;
-    pointer-events: none;
-    overflow: hidden;
-    user-select: none;
-  }
-</style>
